requirements.txt
Django==5.1.6
django_environ=0.12.0
gunicorn=23.0.0
PyYAML==6.0.1
django-mathfilters==1.0.0
regex==2024.11.6
django-esi==6.0.0
whitenoise=6.9.0
